+++ packages/fetch-proxy/src/lib/fetch-proxy.test.ts
@@ -3,55 +3,57 @@ import { describe, it } from 'node:test';
 
 import { type FetchProxyOptions, createFetchProxy } from './fetch-proxy.ts';
 
-async function runProxy(
-  request: Request,
+async function testProxy(
+  input: URL | RequestInfo,
+  init: RequestInit | undefined,
   target: string | URL,
   options?: FetchProxyOptions,
 ): Promise<{ request: Request; response: Response }> {
-  let outgoingRequest: Request;
+  let request: Request;
   let proxy = createFetchProxy(target, {
     ...options,
-    async fetch(input, init) {
-      outgoingRequest = new Request(input, init);
-      return options?.fetch?.(input, init) ?? new Response();
+    fetch(input, init) {
+      request = new Request(input, init);
+      return options?.fetch?.(input, init) ?? Promise.resolve(new Response());
     },
   });
 
-  let proxyResponse = await proxy(request);
+  let response = await proxy(input, init);
 
-  assert.ok(outgoingRequest!);
+  assert.ok(request!);
 
-  return {
-    request: outgoingRequest,
-    response: proxyResponse,
-  };
+  return { request, response };
 }
 
 describe('fetch proxy', () => {
   it('appends the request URL pathname + search to the target URL', async () => {
-    let { request: request1 } = await runProxy(
-      new Request('http://shopify.com'),
+    let { request: request1 } = await testProxy(
+      'http://shopify.com',
+      undefined,
       'https://remix.run:3000/rsc',
     );
 
     assert.equal(request1.url, 'https://remix.run:3000/rsc');
 
-    let { request: request2 } = await runProxy(
-      new Request('http://shopify.com/?q=remix'),
+    let { request: request2 } = await testProxy(
+      'http://shopify.com/?q=remix',
+      undefined,
       'https://remix.run:3000/rsc',
     );
 
     assert.equal(request2.url, 'https://remix.run:3000/rsc?q=remix');
 
-    let { request: request3 } = await runProxy(
-      new Request('http://shopify.com/search?q=remix'),
+    let { request: request3 } = await testProxy(
+      'http://shopify.com/search?q=remix',
+      undefined,
       'https://remix.run:3000/',
     );
 
     assert.equal(request3.url, 'https://remix.run:3000/search?q=remix');
 
-    let { request: request4 } = await runProxy(
-      new Request('http://shopify.com/search?q=remix'),
+    let { request: request4 } = await testProxy(
+      'http://shopify.com/search?q=remix',
+      undefined,
       'https://remix.run:3000/rsc',
     );
 
@@ -59,14 +61,15 @@ describe('fetch proxy', () => {
   });
 
   it('forwards request method, headers, and body', async () => {
-    let { request } = await runProxy(
-      new Request('http://shopify.com/search?q=remix', {
+    let { request } = await testProxy(
+      'http://shopify.com/search?q=remix',
+      {
         method: 'POST',
         headers: {
           'Content-Type': 'text/plain',
         },
         body: 'hello',
-      }),
+      },
       'https://remix.run:3000/rsc',
     );
 
@@ -75,9 +78,24 @@ describe('fetch proxy', () => {
     assert.equal(await request.text(), 'hello');
   });
 
+  it('forwards an empty request body', async () => {
+    let { request } = await testProxy(
+      'http://shopify.com/search?q=remix',
+      {
+        method: 'POST',
+      },
+      'https://remix.run:3000/rsc',
+    );
+
+    assert.equal(request.method, 'POST');
+    assert.equal(request.headers.get('Content-Type'), null);
+    assert.equal(await request.text(), '');
+  });
+
   it('does not append X-Forwarded-Proto and X-Forwarded-Host headers by default', async () => {
-    let { request } = await runProxy(
-      new Request('http://shopify.com:8080/search?q=remix'),
+    let { request } = await testProxy(
+      'http://shopify.com:8080/search?q=remix',
+      undefined,
       'https://remix.run:3000/rsc',
     );
 
@@ -86,8 +104,9 @@ describe('fetch proxy', () => {
   });
 
   it('appends X-Forwarded-Proto and X-Forwarded-Host headers when desired', async () => {
-    let { request } = await runProxy(
-      new Request('http://shopify.com:8080/search?q=remix'),
+    let { request } = await testProxy(
+      'http://shopify.com:8080/search?q=remix',
+      undefined,
       'https://remix.run:3000/rsc',
       {
         xForwardedHeaders: true,
@@ -98,9 +117,26 @@ describe('fetch proxy', () => {
     assert.equal(request.headers.get('X-Forwarded-Host'), 'shopify.com:8080');
   });
 
+  it('forwards additional request init options', async () => {
+    let { request } = await testProxy(
+      'http://shopify.com/search?q=remix',
+      {
+        cache: 'no-cache',
+        credentials: 'include',
+        redirect: 'manual',
+      },
+      'https://remix.run:3000/rsc',
+    );
+
+    assert.equal(request.cache, 'no-cache');
+    assert.equal(request.credentials, 'include');
+    assert.equal(request.redirect, 'manual');
+  });
+
   it('rewrites cookie domain and path', async () => {
-    let { response } = await runProxy(
-      new Request('http://shopify.com/search?q=remix'),
+    let { response } = await testProxy(
+      'http://shopify.com/search?q=remix',
+      undefined,
       'https://remix.run:3000/rsc',
       {
         async fetch() {
@@ -122,8 +158,9 @@ describe('fetch proxy', () => {
   });
 
   it('does not rewrite cookie domain and path when opting-out', async () => {
-    let { response } = await runProxy(
-      new Request('http://shopify.com/?q=remix'),
+    let { response } = await testProxy(
+      'http://shopify.com/?q=remix',
+      undefined,
       'https://remix.run:3000/rsc',
       {
         rewriteCookieDomain: false,